+++ pkg/pool-utils/contracts/test/oracle/MockPoolPriceOracle.sol
@@ -22,6 +22,7 @@ import "../../interfaces/IPriceOracle.sol";
 import "./MockSamples.sol";
 
 contract MockPoolPriceOracle is MockSamples, PoolPriceOracle {
+    using Buffer for uint256;
     using Samples for bytes32;
 
     struct BinarySearchResult {
@@ -74,9 +75,14 @@ contract MockPoolPriceOracle is MockSamples, PoolPriceOracle {
         view
         returns (BinarySearchResult[] memory results)
     {
+        uint256 oldestIndex = _mockedOracleIndex.next();
+        bytes32 oldestSample = _getSample(oldestIndex);
+        uint256 oldestTimestamp = oldestSample.timestamp();
+        uint256 length = oldestTimestamp > 0 ? Buffer.SIZE : oldestIndex;
+
         results = new BinarySearchResult[](dates.length);
         for (uint256 i = 0; i < dates.length; i++) {
-            (bytes32 prev, bytes32 next) = _findNearestSample(dates[i], offset);
+            (bytes32 prev, bytes32 next) = _findNearestSample(dates[i], offset, length);
             results[i] = BinarySearchResult({ prev: prev.timestamp(), next: next.timestamp() });
         }
     }